+++ stl/lib/stl.dart
@@ -49,3 +49,4 @@ export 'src/primitives/uint32.dart';
 export 'src/primitives/uint64.dart';
 export 'src/math/constant/constant.dart';
 export 'src/math/numeric.dart';
+export 'src/math/complex.dart';